Split agent-card.tsx into header, status, details, and actions files
agent-card.tsx had grown to 829 lines rendering four logically independent
regions in one render tree. Each region is now its own file, leaving the
card as a composition shell that owns only layout, expansion state, and
the drag-handler effect.

- agent-card-header.tsx (254) — type icon, session name, rename prompt
  (state and handler moved along with it), status badges, resume, expand
  toggle
- agent-card-status.tsx (160) — setup/archive phase lines, latest-event
  summary, repo label
- agent-card-details.tsx (207) — branch/worktree panel, diff stats, IDE
  launch, sandbox indicator (owns its own copy-to-clipboard state)
- agent-card-actions.tsx (123) — persona launcher plus lifecycle controls
- agent-card.tsx 829 → 267

Purely structural: markup, class names, and test IDs are unchanged.
